add queue method

+++ roomie/src/main/java/com/pm/roomie/dao/FlatMemberRepository.java
@@ -2,7 +2,7 @@ package com.pm.roomie.dao;
 
 import com.pm.roomie.json.Flat;
 import com.pm.roomie.json.FlatMember;
-import com.pm.roomie.json.Flat;
+
 import com.pm.roomie.json.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FlatMemberRepository extends JpaRepository<FlatMember,Long> {
+public interface FlatMemberRepository extends JpaRepository<FlatMember,Integer> {
 
     List<FlatMember> findByUser(User user);
     List<FlatMember> findByFlat(Flat flat);
